+++ tests/test_upload.py
@@ -53,11 +53,11 @@ class Test(CLITestCase):
         content = {'package_types': ['conda']}
         registry.register(method='GET', path='/package/eggs/foo', content=content)
         registry.register(method='GET', path='/release/eggs/foo/0.1', content='{}')
-        registry.register(method='GET', path='/dist/eggs/foo/0.1/osx-64/foo-0.1-0.tar.bz2', status=404, content='{}')
 
         content = {'post_url': 'http://s3url.com/s3_url', 'form_data': {}, 'dist_id': 'dist_id'}
-        staging_response = registry.register(method='POST', path='/stage/eggs/foo/0.1/osx-64/foo-0.1-0.tar.bz2',
-                                             content=content)
+        staging_response = registry.register(
+            method='POST', path='/stage/eggs/foo/0.1/osx-64/foo-0.1-0.tar.bz2', content=content,
+        )
 
         registry.register(method='POST', path='/s3_url', status=201)
         registry.register(method='POST', path='/commit/eggs/foo/0.1/osx-64/foo-0.1-0.tar.bz2', status=200, content={})
@@ -74,16 +74,16 @@ class Test(CLITestCase):
         content = {'package_types': ['conda']}
         registry.register(method='GET', path='/package/eggs/mock', content=content)
         registry.register(method='GET', path='/release/eggs/mock/2.0.0', content='{}')
-        registry.register(
-            method='GET', path='/dist/eggs/mock/2.0.0/osx-64/mock-2.0.0-py37_1000.conda', status=404, content='{}')
 
         content = {'post_url': 'http://s3url.com/s3_url', 'form_data': {}, 'dist_id': 'dist_id'}
         staging_response = registry.register(
-            method='POST', path='/stage/eggs/mock/2.0.0/osx-64/mock-2.0.0-py37_1000.conda', content=content)
+            method='POST', path='/stage/eggs/mock/2.0.0/osx-64/mock-2.0.0-py37_1000.conda', content=content,
+        )
 
         registry.register(method='POST', path='/s3_url', status=201)
         registry.register(
-            method='POST', path='/commit/eggs/mock/2.0.0/osx-64/mock-2.0.0-py37_1000.conda', status=200, content={})
+            method='POST', path='/commit/eggs/mock/2.0.0/osx-64/mock-2.0.0-py37_1000.conda', status=200, content={},
+        )
 
         main(['--show-traceback', 'upload', data_dir('mock-2.0.0-py37_1000.conda')], False)
 
@@ -98,16 +98,16 @@ class Test(CLITestCase):
         registry.register(method='GET', path='/package/eggs/mock', content=content)
         registry.register(method='GET', path='/release/eggs/mock/2.0.0', content='{}')
         registry.register(method='PATCH', path='/release/eggs/mock/2.0.0', content='{}')
-        registry.register(
-            method='GET', path='/dist/eggs/mock/2.0.0/osx-64/mock-2.0.0-py37_1000.conda', status=404, content='{}')
 
         content = {'post_url': 'http://s3url.com/s3_url', 'form_data': {}, 'dist_id': 'dist_id'}
         staging_response = registry.register(
-            method='POST', path='/stage/eggs/mock/2.0.0/osx-64/mock-2.0.0-py37_1000.conda', content=content)
+            method='POST', path='/stage/eggs/mock/2.0.0/osx-64/mock-2.0.0-py37_1000.conda', content=content,
+        )
 
         registry.register(method='POST', path='/s3_url', status=201)
         registry.register(
-            method='POST', path='/commit/eggs/mock/2.0.0/osx-64/mock-2.0.0-py37_1000.conda', status=200, content={})
+            method='POST', path='/commit/eggs/mock/2.0.0/osx-64/mock-2.0.0-py37_1000.conda', status=200, content={},
+        )
 
         main(['--show-traceback', 'upload', '--force-metadata-update', data_dir('mock-2.0.0-py37_1000.conda')], False)
 
@@ -121,17 +121,16 @@ class Test(CLITestCase):
         content = {'package_types': ['pypi']}
         registry.register(method='GET', path='/package/eggs/test-package34', content=content)
         registry.register(method='GET', path='/release/eggs/test-package34/0.3.1', content='{}')
-        registry.register(method='GET', path='/dist/eggs/test-package34/0.3.1/test_package34-0.3.1.tar.gz', status=404,
-                          content='{}')
 
         content = {'post_url': 'http://s3url.com/s3_url', 'form_data': {}, 'dist_id': 'dist_id'}
         staging_response = registry.register(
-            method='POST', path='/stage/eggs/test-package34/0.3.1/test_package34-0.3.1.tar.gz',
-            content=content)
+            method='POST', path='/stage/eggs/test-package34/0.3.1/test_package34-0.3.1.tar.gz', content=content,
+        )
 
         registry.register(method='POST', path='/s3_url', status=201)
-        registry.register(method='POST', path='/commit/eggs/test-package34/0.3.1/test_package34-0.3.1.tar.gz',
-                          status=200, content={})
+        registry.register(
+            method='POST', path='/commit/eggs/test-package34/0.3.1/test_package34-0.3.1.tar.gz', status=200, content={},
+        )
 
         main(['--show-traceback', 'upload', data_dir('test_package34-0.3.1.tar.gz')], False)
 
@@ -145,17 +144,16 @@ class Test(CLITestCase):
         content = {'package_types': ['pypi']}
         registry.register(method='GET', path='/package/eggs/test_package34', content=content)
         registry.register(method='GET', path='/release/eggs/test_package34/0.3.1', content='{}')
-        registry.register(method='GET', path='/dist/eggs/test_package34/0.3.1/test_package34-0.3.1.tar.gz', status=404,
-                          content='{}')
 
         content = {'post_url': 'http://s3url.com/s3_url', 'form_data': {}, 'dist_id': 'dist_id'}
         staging_response = registry.register(
-            method='POST', path='/stage/eggs/test_package34/0.3.1/test_package34-0.3.1.tar.gz',
-            content=content)
+            method='POST', path='/stage/eggs/test_package34/0.3.1/test_package34-0.3.1.tar.gz', content=content,
+        )
 
         registry.register(method='POST', path='/s3_url', status=201)
-        registry.register(method='POST', path='/commit/eggs/test_package34/0.3.1/test_package34-0.3.1.tar.gz',
-                          status=200, content={})
+        registry.register(
+            method='POST', path='/commit/eggs/test_package34/0.3.1/test_package34-0.3.1.tar.gz', status=200, content={},
+        )
 
         # Pass -o to override the channel/package pypi package should go to
         main(['--show-traceback', 'upload',
@@ -172,10 +170,10 @@ class Test(CLITestCase):
 
         # Passing -o for `file` package_type doesn't override channel
         with self.assertRaises(errors.BinstarError):
-            main(['--show-traceback', 'upload',
-                  '--package', 'test_package',
-                  '--package-type', 'file',
-                  data_dir('test_package34-0.3.1.tar.gz')], False)
+            main([
+                '--show-traceback', 'upload', '--package', 'test_package', '--package-type', 'file',
+                data_dir('test_package34-0.3.1.tar.gz'),
+            ], False)
 
         registry.assertAllCalled()
 
@@ -185,9 +183,9 @@ class Test(CLITestCase):
         registry.register(method='GET', path='/user', content='{"login": "eggs"}')
 
         with self.assertRaises(errors.BinstarError):
-            main(['--show-traceback', 'upload',
-                  '--package', 'alpha_omega',
-                  data_dir('test_package34-0.3.1.tar.gz')], False)
+            main([
+                '--show-traceback', 'upload', '--package', 'alpha_omega', data_dir('test_package34-0.3.1.tar.gz'),
+            ], False)
 
         registry.assertAllCalled()
 
@@ -198,22 +196,21 @@ class Test(CLITestCase):
         content = {'package_types': ['file']}
         registry.register(method='GET', path='/package/eggs/test-package34', content=content)
         registry.register(method='GET', path='/release/eggs/test-package34/0.3.1', content='{}')
-        registry.register(method='GET', path='/dist/eggs/test-package34/0.3.1/test_package34-0.3.1.tar.gz', status=404,
-                          content='{}')
 
         content = {'post_url': 'http://s3url.com/s3_url', 'form_data': {}, 'dist_id': 'dist_id'}
         staging_response = registry.register(
-            method='POST', path='/stage/eggs/test-package34/0.3.1/test_package34-0.3.1.tar.gz', content=content)
+            method='POST', path='/stage/eggs/test-package34/0.3.1/test_package34-0.3.1.tar.gz', content=content,
+        )
 
         registry.register(method='POST', path='/s3_url', status=201)
-        registry.register(method='POST', path='/commit/eggs/test-package34/0.3.1/test_package34-0.3.1.tar.gz',
-                          status=200, content={})
+        registry.register(
+            method='POST', path='/commit/eggs/test-package34/0.3.1/test_package34-0.3.1.tar.gz', status=200, content={},
+        )
 
-        main(['--show-traceback', 'upload',
-              '--package-type', 'file',
-              '--package', 'test-package34',
-              '--version', '0.3.1',
-              data_dir('test_package34-0.3.1.tar.gz')], False)
+        main([
+            '--show-traceback', 'upload', '--package-type', 'file', '--package', 'test-package34', '--version', '0.3.1',
+            data_dir('test_package34-0.3.1.tar.gz'),
+        ], False)
 
         registry.assertAllCalled()
         self.assertIsNotNone(json.loads(staging_response.req.body).get('sha256'))
@@ -228,14 +225,11 @@ class Test(CLITestCase):
         registry.register(method='GET', path='/user', content='{"login": "eggs"}')
         registry.register(method='GET', path='/apps/eggs/projects/dog', content='{}')
         stage_content = '{"post_url":"http://s3url.com/s3_url", "form_data":{"foo":"bar"}, "dist_id":"dist42"}'
-        registry.register(method='POST', path='/apps/eggs/projects/dog/stage',
-                          content=stage_content)
+        registry.register(method='POST', path='/apps/eggs/projects/dog/stage', content=stage_content)
         registry.register(method='POST', path='/s3_url', status=201)
         registry.register(method='POST', path='/apps/eggs/projects/dog/commit/dist42', content='{}')
 
-        main(['--show-traceback', 'upload',
-              '--package-type', 'project',
-              data_dir('bar')], False)
+        main(['--show-traceback', 'upload', '--package-type', 'project', data_dir('bar')], False)
 
         registry.assertAllCalled()
 
@@ -246,14 +240,11 @@ class Test(CLITestCase):
         registry.register(method='GET', path='/user', content='{"login": "eggs"}')
         registry.register(method='GET', path='/apps/eggs/projects/foo', content='{}')
         stage_content = '{"post_url":"http://s3url.com/s3_url", "form_data":{"foo":"bar"}, "dist_id":"dist42"}'
-        registry.register(method='POST', path='/apps/eggs/projects/foo/stage',
-                          content=stage_content)
+        registry.register(method='POST', path='/apps/eggs/projects/foo/stage', content=stage_content)
         registry.register(method='POST', path='/s3_url', status=201)
         registry.register(method='POST', path='/apps/eggs/projects/foo/commit/dist42', content='{}')
 
-        main(['--show-traceback', 'upload',
-              '--package-type', 'project',
-              data_dir('foo.ipynb')], False)
+        main(['--show-traceback', 'upload', '--package-type', 'project', data_dir('foo.ipynb')], False)
 
         registry.assertAllCalled()
 
@@ -267,16 +258,16 @@ class Test(CLITestCase):
         content = {'package_types': ['ipynb']}
         registry.register(method='GET', path='/package/eggs/foo', content=content)
         registry.register(method='GET', path='/release/eggs/foo/{}'.format(mock_version), content='{}')
-        registry.register(
-            method='GET', path='/dist/eggs/foo/{}/foo.ipynb'.format(mock_version), status=404, content='{}')
 
         content = {'post_url': 'http://s3url.com/s3_url', 'form_data': {}, 'dist_id': 'dist_id'}
-        staging_response = registry.register(method='POST', path='/stage/eggs/foo/{}/foo.ipynb'.format(mock_version),
-                                             content=content)
+        staging_response = registry.register(
+            method='POST', path='/stage/eggs/foo/{}/foo.ipynb'.format(mock_version), content=content,
+        )
 
         registry.register(method='POST', path='/s3_url', status=201)
         registry.register(
-            method='POST', path='/commit/eggs/foo/{}/foo.ipynb'.format(mock_version), status=200, content={})
+            method='POST', path='/commit/eggs/foo/{}/foo.ipynb'.format(mock_version), status=200, content={},
+        )
 
         with patch('binstar_client.inspect_package.ipynb.datetime') as mock_datetime:
             mock_datetime.now.return_value = test_datetime
@@ -291,35 +282,29 @@ class Test(CLITestCase):
         registry.register(method='GET', path='/user/alice', content='{"login": "alice"}')
         registry.register(method='GET', path='/apps/alice/projects/dog', content='{}')
         stage_content = '{"post_url":"http://s3url.com/s3_url", "form_data":{"foo":"bar"}, "dist_id":"dist42"}'
-        registry.register(method='POST', path='/apps/alice/projects/dog/stage',
-                          content=stage_content)
+        registry.register(method='POST', path='/apps/alice/projects/dog/stage', content=stage_content)
         registry.register(method='POST', path='/s3_url', status=201)
         registry.register(method='POST', path='/apps/alice/projects/dog/commit/dist42', content='{}')
 
-        main(['--show-traceback', 'upload',
-              '--package-type', 'project',
-              '--user', 'alice',
-              data_dir('bar')], False)
+        main(['--show-traceback', 'upload', '--package-type', 'project', '--user', 'alice', data_dir('bar')], False)
 
         registry.assertAllCalled()
 
     @urlpatch
     def test_upload_project_specifying_token(self, registry):
         registry.register(method='HEAD', path='/', status=200)
-        registry.register(method='GET', path='/user/eggs', content='{"login": "eggs"}',
-                          expected_headers={'Authorization': 'token abcdefg'})
+        registry.register(
+            method='GET', path='/user/eggs', content='{"login": "eggs"}',
+            expected_headers={'Authorization': 'token abcdefg'},
+        )
         registry.register(method='GET', path='/user', content='{"login": "eggs"}')
         registry.register(method='GET', path='/apps/eggs/projects/dog', content='{}')
         stage_content = '{"post_url":"http://s3url.com/s3_url", "form_data":{"foo":"bar"}, "dist_id":"dist42"}'
-        registry.register(method='POST', path='/apps/eggs/projects/dog/stage',
-                          content=stage_content)
+        registry.register(method='POST', path='/apps/eggs/projects/dog/stage', content=stage_content)
         registry.register(method='POST', path='/s3_url', status=201)
         registry.register(method='POST', path='/apps/eggs/projects/dog/commit/dist42', content='{}')
 
-        main(['--show-traceback', '--token', 'abcdefg',
-              'upload',
-              '--package-type', 'project',
-              data_dir('bar')], False)
+        main(['--show-traceback', '--token', 'abcdefg', 'upload', '--package-type', 'project', data_dir('bar')], False)
 
         registry.assertAllCalled()
 
@@ -332,12 +317,50 @@ class Test(CLITestCase):
         content = {'package_types': ['conda']}
         registry.register(method='GET', path='/package/eggs/foo', content=content)
         registry.register(method='GET', path='/release/eggs/foo/0.1', content='{}')
-        registry.register(method='GET', path='/dist/eggs/foo/0.1/osx-64/foo-0.1-0.tar.bz2', status=200, content='{}')
+        query_001 = registry.register(method='POST', path='/stage/eggs/foo/0.1/osx-64/foo-0.1-0.tar.bz2', status=409)
 
-        # don't overwrite
-        bool_input.return_value = False
+        bool_input.return_value = False  # do not overwrite package
 
         main(['--show-traceback', 'upload', '-i', data_dir('foo-0.1-0.tar.bz2')], False)
+
+        registry.assertAllCalled()
+        self.assertIsNotNone(json.loads(query_001.req.body).get('sha256'))
+
+    @urlpatch
+    @patch('binstar_client.commands.upload.bool_input')
+    def test_upload_interactive_overwrite(self, registry, bool_input):
+        registry.register(method='HEAD', path='/', status=200)
+        registry.register(method='GET', path='/user', content='{"login": "eggs"}')
+        content = {'package_types': ['conda']}
+        registry.register(method='GET', path='/package/eggs/foo', content=content)
+        registry.register(method='GET', path='/release/eggs/foo/0.1', content='{}')
+        query_001 = registry.register(method='POST', path='/stage/eggs/foo/0.1/osx-64/foo-0.1-0.tar.bz2', status=409)
+        query_002 = None
+
+        def allow_upload():
+            nonlocal query_002
+            query_002 = registry.register(
+                method='POST', path='/stage/eggs/foo/0.1/osx-64/foo-0.1-0.tar.bz2',
+                content={'post_url': 'http://s3url.com/s3_url', 'form_data': {}, 'dist_id': 'dist_id'},
+            )
+            registry.register(method='POST', path='/s3_url', status=201)
+            registry.register(
+                method='POST', path='/commit/eggs/foo/0.1/osx-64/foo-0.1-0.tar.bz2', status=200, content={},
+            )
+
+        registry.register(
+            method='DELETE', path='/dist/eggs/foo/0.1/osx-64/foo-0.1-0.tar.bz2', content='{}',
+            side_effect=allow_upload,
+        )
+
+        bool_input.return_value = True
+
+        main(['--show-traceback', 'upload', '-i', data_dir('foo-0.1-0.tar.bz2')], False)
+
+        registry.assertAllCalled()
+        self.assertIsNotNone(json.loads(query_001.req.body).get('sha256'))
+        self.assertIsNotNone(query_002)
+        self.assertIsNotNone(json.loads(query_002.req.body).get('sha256'))
 
     @urlpatch
     def test_upload_private_package(self, registry):
@@ -347,11 +370,11 @@ class Test(CLITestCase):
         content = {'package_types': ['conda']}
         registry.register(method='POST', path='/package/eggs/foo', content=content, status=200)
         registry.register(method='GET', path='/release/eggs/foo/0.1', content='{}')
-        registry.register(method='GET', path='/dist/eggs/foo/0.1/osx-64/foo-0.1-0.tar.bz2', status=404, content='{}')
 
         content = {'post_url': 'http://s3url.com/s3_url', 'form_data': {}, 'dist_id': 'dist_id'}
         staging_response = registry.register(
-            method='POST', path='/stage/eggs/foo/0.1/osx-64/foo-0.1-0.tar.bz2', content=content)
+            method='POST', path='/stage/eggs/foo/0.1/osx-64/foo-0.1-0.tar.bz2', content=content,
+        )
 
         registry.register(method='POST', path='/s3_url', status=201)
         registry.register(method='POST', path='/commit/eggs/foo/0.1/osx-64/foo-0.1-0.tar.bz2', status=200, content={})
@@ -366,8 +389,10 @@ class Test(CLITestCase):
         registry.register(method='HEAD', path='/', status=200)
         registry.register(method='GET', path='/user', content='{"login": "eggs"}')
         registry.register(method='GET', path='/package/eggs/foo', content='{}', status=404)
-        registry.register(method='POST', path='/package/eggs/foo',
-                          content='{"error": "You can not create a private package."}', status=400)
+        registry.register(
+            method='POST', path='/package/eggs/foo', content='{"error": "You can not create a private package."}',
+            status=400,
+        )
 
         with self.assertRaises(errors.BinstarError):
             main(['--show-traceback', 'upload', '--private', data_dir('foo-0.1-0.tar.bz2')], False)